modules/contract_analyzer.py
from web3 import Web3
from typing import Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

class ContractAnalyzer:
    def __init__(self):
        self.w3 = Web3(Web3.HTTPProvider(os.getenv('ETHEREUM_RPC_URL')))

    async def analyze(self, token_address: str) -> Dict[str, Any]:
        """
        Analyze a token's smart contract.
        """
        try:
            # Validate address
            if not self.w3.is_address(token_address):
                raise ValueError("Invalid token address")

            # Get contract data
            contract_data = await self._get_contract_data(token_address)
            
            # Analyze contract features
            features = self._analyze_features(contract_data)
            
            # Check for known issues
            issues = self._check_issues(contract_data)

            return {
                'address': token_address,
                'type': contract_data.get('type', 'Unknown'),
                'features': features,
                'known_issues': issues,
                'contract_data': contract_data
            }
        except Exception as e:
            print(f"Error in contract analysis: {str(e)}")
            return {
                'address': token_address,
                'type': 'Unknown',
                'features': [],
                'known_issues': [f"Error analyzing contract: {str(e)}"],
                'contract_data': {}
            }

    async def _get_contract_data(self, token_address: str) -> Dict[str, Any]:
        """
        Get contract data and metadata.
        """
        try:
            # This is a placeholder. You would need to implement actual contract data fetching
            # using Web3.py and contract ABIs
            return {
                'type': 'ERC20',
                'name': 'Unknown',
                'symbol': 'UNKNOWN',
                'decimals': 18,
                'total_supply': 0,
                'owner': '0x0000000000000000000000000000000000000000',
                'verified': False,
                'source_code': '',
                'abi': []
            }
        except Exception:
            return {
                'type': 'Unknown',
                'name': 'Unknown',
                'symbol': 'UNKNOWN',
                'decimals': 18,
                'total_supply': 0,
                'owner': '0x0000000000000000000000000000000000000000',
                'verified': False,
                'source_code': '',
                'abi': []
            }

    def _analyze_features(self, contract_data: Dict[str, Any]) -> list:
        """
        Analyze contract features and capabilities.
        """
        features = []
        
        # Check if contract is verified
        if contract_data.get('verified', False):
            features.append('Verified Contract')
        
        # Check for common features
        if contract_data.get('type') == 'ERC20':
            features.append('ERC20 Standard')
        
        # Add more feature checks here
        
        return features

    def _check_issues(self, contract_data: Dict[str, Any]) -> list:
        """
        Check for known issues and vulnerabilities.
        """
        issues = []
        
        # Check if contract is verified
        if not contract_data.get('verified', False):
            issues.append('Unverified Contract')
        
        # Check for common issues
        if contract_data.get('owner') == '0x0000000000000000000000000000000000000000':
            issues.append('No Owner Address')
        
        # Add more issue checks here
        
        return issues 