requirements.txt
python-telegram-bot==20.7
web3==6.11.1
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
textblob==0.17.1
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
aiohttp==3.9.1
asyncio==3.4.3
wheel==0.41.2
setuptools==68.2.2
Pillow==10.1.0
google-generativeai==0.3.2 